requirements.txt
phidata
yfinance
packaging
fastapi
uvicorn
groq
openai
python - docx
